Move _MK and _CV implementation functions to separate cpp
