+++ internal/c/libqb.cpp
@@ -6847,7 +6847,6 @@ int32 lock_display_required = 0;
 #define cost_delay 0
 uint32 cost = 0;
 
-#include "msbin.c"
 
 int64 build_int64(uint32 val2, uint32 val1) {
     static int64 val;
@@ -7200,276 +7199,6 @@ qbs *func_varptr_helper(uint8 type, uint16 offset) {
     tqbs->chr[1] = offset & 255;
     tqbs->chr[2] = offset >> 8;
     return tqbs;
-}
-
-qbs *func_mksmbf(float val) {
-    static qbs *tqbs;
-    tqbs = qbs_new(4, 1);
-    if (_fieeetomsbin(&val, (float *)tqbs->chr) == 1) {
-        error(5);
-        tqbs->len = 0;
-    }
-    return tqbs;
-}
-qbs *func_mkdmbf(double val) {
-    static qbs *tqbs;
-    tqbs = qbs_new(8, 1);
-    if (_dieeetomsbin(&val, (double *)tqbs->chr) == 1) {
-        error(5);
-        tqbs->len = 0;
-    }
-    return tqbs;
-}
-
-float func_cvsmbf(qbs *str) {
-    static float val;
-    if (str->len < 4) {
-        error(5);
-        return 0;
-    }
-    if (_fmsbintoieee((float *)str->chr, &val) == 1) {
-        error(5);
-        return 0;
-    }
-    return val;
-}
-double func_cvdmbf(qbs *str) {
-    static double val;
-    if (str->len < 8) {
-        error(5);
-        return 0;
-    }
-    if (_dmsbintoieee((double *)str->chr, &val) == 1) {
-        error(5);
-        return 0;
-    }
-    return val;
-}
-
-qbs *b2string(char v) {
-    static qbs *tqbs;
-    tqbs = qbs_new(1, 1);
-    *((char *)(tqbs->chr)) = v;
-    return tqbs;
-}
-qbs *ub2string(char v) {
-    static qbs *tqbs;
-    tqbs = qbs_new(1, 1);
-    *((uint8 *)(tqbs->chr)) = v;
-    return tqbs;
-}
-qbs *i2string(int16 v) {
-    static qbs *tqbs;
-    tqbs = qbs_new(2, 1);
-    *((int16 *)(tqbs->chr)) = v;
-    return tqbs;
-}
-qbs *ui2string(int16 v) {
-    static qbs *tqbs;
-    tqbs = qbs_new(2, 1);
-    *((uint16 *)(tqbs->chr)) = v;
-    return tqbs;
-}
-qbs *l2string(int32 v) {
-    static qbs *tqbs;
-    tqbs = qbs_new(4, 1);
-    *((int32 *)(tqbs->chr)) = v;
-    return tqbs;
-}
-qbs *ul2string(uint32 v) {
-    static qbs *tqbs;
-    tqbs = qbs_new(4, 1);
-    *((uint32 *)(tqbs->chr)) = v;
-    return tqbs;
-}
-qbs *i642string(int64 v) {
-    static qbs *tqbs;
-    tqbs = qbs_new(8, 1);
-    *((int64 *)(tqbs->chr)) = v;
-    return tqbs;
-}
-qbs *ui642string(uint64 v) {
-    static qbs *tqbs;
-    tqbs = qbs_new(8, 1);
-    *((uint64 *)(tqbs->chr)) = v;
-    return tqbs;
-}
-qbs *s2string(float v) {
-    static qbs *tqbs;
-    tqbs = qbs_new(4, 1);
-    *((float *)(tqbs->chr)) = v;
-    return tqbs;
-}
-qbs *d2string(double v) {
-    static qbs *tqbs;
-    tqbs = qbs_new(8, 1);
-    *((double *)(tqbs->chr)) = v;
-    return tqbs;
-}
-qbs *f2string(long double v) {
-    static qbs *tqbs;
-    tqbs = qbs_new(32, 1);
-    memset(tqbs->chr, 0, 32);
-    *((long double *)(tqbs->chr)) = v;
-    return tqbs;
-}
-qbs *o2string(ptrszint v) {
-    static qbs *tqbs;
-    tqbs = qbs_new(sizeof(ptrszint), 1);
-    memset(tqbs->chr, 0, sizeof(ptrszint));
-    *((ptrszint *)(tqbs->chr)) = v;
-    return tqbs;
-}
-qbs *uo2string(uptrszint v) {
-    static qbs *tqbs;
-    tqbs = qbs_new(sizeof(uptrszint), 1);
-    memset(tqbs->chr, 0, sizeof(uptrszint));
-    *((uptrszint *)(tqbs->chr)) = v;
-    return tqbs;
-}
-qbs *bit2string(uint32 bsize, int64 v) {
-    static qbs *tqbs;
-    tqbs = qbs_new(8, 1);
-    int64 bmask;
-    bmask = ~(-(((int64)1) << bsize));
-    *((int64 *)(tqbs->chr)) = v & bmask;
-    tqbs->len = (bsize + 7) >> 3;
-    return tqbs;
-}
-qbs *ubit2string(uint32 bsize, uint64 v) {
-    static qbs *tqbs;
-    int64 bmask;
-    tqbs = qbs_new(8, 1);
-    bmask = ~(-(((int64)1) << bsize));
-    *((uint64 *)(tqbs->chr)) = v & bmask;
-    tqbs->len = (bsize + 7) >> 3;
-    return tqbs;
-}
-
-char string2b(qbs *str) {
-    if (str->len < 1) {
-        error(5);
-        return 0;
-    } else {
-        return *((char *)str->chr);
-    }
-}
-uint8 string2ub(qbs *str) {
-    if (str->len < 1) {
-        error(5);
-        return 0;
-    } else {
-        return *((uint8 *)str->chr);
-    }
-}
-int16 string2i(qbs *str) {
-    if (str->len < 2) {
-        error(5);
-        return 0;
-    } else {
-        return *((int16 *)str->chr);
-    }
-}
-uint16 string2ui(qbs *str) {
-    if (str->len < 2) {
-        error(5);
-        return 0;
-    } else {
-        return *((uint16 *)str->chr);
-    }
-}
-int32 string2l(qbs *str) {
-    if (str->len < 4) {
-        error(5);
-        return 0;
-    } else {
-        return *((int32 *)str->chr);
-    }
-}
-uint32 string2ul(qbs *str) {
-    if (str->len < 4) {
-        error(5);
-        return 0;
-    } else {
-        return *((uint32 *)str->chr);
-    }
-}
-int64 string2i64(qbs *str) {
-    if (str->len < 8) {
-        error(5);
-        return 0;
-    } else {
-        return *((int64 *)str->chr);
-    }
-}
-uint64 string2ui64(qbs *str) {
-    if (str->len < 8) {
-        error(5);
-        return 0;
-    } else {
-        return *((uint64 *)str->chr);
-    }
-}
-float string2s(qbs *str) {
-    if (str->len < 4) {
-        error(5);
-        return 0;
-    } else {
-        return *((float *)str->chr);
-    }
-}
-double string2d(qbs *str) {
-    if (str->len < 8) {
-        error(5);
-        return 0;
-    } else {
-        return *((double *)str->chr);
-    }
-}
-long double string2f(qbs *str) {
-    if (str->len < 32) {
-        error(5);
-        return 0;
-    } else {
-        return *((long double *)str->chr);
-    }
-}
-ptrszint string2o(qbs *str) {
-    if (str->len < sizeof(ptrszint)) {
-        error(5);
-        return 0;
-    } else {
-        return *((ptrszint *)str->chr);
-    }
-}
-uptrszint string2uo(qbs *str) {
-    if (str->len < sizeof(uptrszint)) {
-        error(5);
-        return 0;
-    } else {
-        return *((uptrszint *)str->chr);
-    }
-}
-uint64 string2ubit(qbs *str, uint32 bsize) {
-    int64 bmask;
-    if (str->len < ((bsize + 7) >> 3)) {
-        error(5);
-        return 0;
-    }
-    bmask = ~(-(((int64)1) << bsize));
-    return (*(uint64 *)str->chr) & bmask;
-}
-int64 string2bit(qbs *str, uint32 bsize) {
-    int64 bmask, bval64;
-    if (str->len < ((bsize + 7) >> 3)) {
-        error(5);
-        return 0;
-    }
-    bmask = ~(-(((int64)1) << bsize));
-    bval64 = (*(uint64 *)str->chr) & bmask;
-    if (bval64 & (((int64)1) << (bsize - 1)))
-        return (bval64 | (~bmask));
-    return bval64;
 }
 
 qbs *qbs_inkey() {
