+++ internal/c/libqb/build.mk
@@ -13,6 +13,7 @@ libqb-objs-y += $(PATH_LIBQB)/src/rounding.o
 libqb-objs-y += $(PATH_LIBQB)/src/qbs.o
 libqb-objs-y += $(PATH_LIBQB)/src/qbs_str.o
 libqb-objs-y += $(PATH_LIBQB)/src/qbs_cmem.o
+libqb-objs-y += $(PATH_LIBQB)/src/qbs_mk_cv.o
 libqb-objs-y += $(PATH_LIBQB)/src/string_functions.o
 
 libqb-objs-$(DEP_HTTP) += $(PATH_LIBQB)/src/http.o
